requirements.txt
aiofiles==0.8.0
altgraph==0.17.2
androidviewclient==20.7.1
Appium-Python-Client==2.2.0
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
culebratester-client==2.0.23
cycler==0.11.0
fb-idb==1.1.7
fonttools==4.32.0
future==0.18.2
grpclib==0.4.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
kiwisolver==1.4.2
macholib==1.16
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.3
outcome==1.1.0
packaging==21.3
Pillow==9.1.0
protobuf==3.20.0
pure-python-adb==0.3.0.dev0
pycodestyle==2.8.0
pycparser==2.21
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
treelib==1.6.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
